InferenceController and SamplingController
